refactor so all tracers can provide TraceFrames
