+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/CallTracerResultConverter.java
@@ -20,8 +20,8 @@ import static org.hyperledger.besu.evm.internal.Words.toAddress;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.processor.TransactionTrace;
 import org.hyperledger.besu.ethereum.core.Transaction;
-import org.hyperledger.besu.ethereum.debug.TraceFrame;
 import org.hyperledger.besu.ethereum.processing.TransactionProcessingResult;
+import org.hyperledger.besu.evm.tracing.TraceFrame;
 
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
