+++ evm/src/main/java/org/hyperledger/besu/evm/tracing/OperationTracer.java
@@ -147,4 +147,14 @@ public interface OperationTracer {
   default boolean isExtendedTracing() {
     return false;
   }
+
+  /**
+   * Get the trace frames collected during execution. This is primarily used for debug tracing
+   * operations.
+   *
+   * @return the list of trace frames, empty by default
+   */
+  default java.util.List<TraceFrame> getTraceFrames() {
+    return java.util.Collections.emptyList();
+  }
 }
